Change CEL type to interface
The cel package now supports both PCRs and Confidential Computing MR
indexes. We currently only allow one type of index per-log, so we
enforce the MR index type is consistent across a log.

This requires users to specify the measurement register type at creation
time of the CEL through a constructor and removes direct access to the
CEL object.